Avoid comparison of unsigned expression
comparison of unsigned expression < 0 is always false

+++ src/ibuskeymap.c
@@ -119,7 +119,7 @@ ibus_keymap_parse_line (gchar  *str,
     if (keycode == 0 && p1 == p2)
         return FALSE;
 
-    if (keycode < 0 || keycode > 255)
+    if ((int) keycode < 0 || keycode > 255)
         return FALSE;
 
     p1 = p2;
